+++ CHANGELOG.md
@@ -22,6 +22,9 @@ All Notable changes to `Backpack CRUD` will be documented in this file
 
 ## [3.5.14] - 2019-02-xx
 
+### Added
+- #1749 - ```$crud->getFilter('name')``` and ```$crud->hasActiveFilter('name')``` methods for Filters;
+
 ### Fixed
 - #1792 - javascript error on Show page, due to Clone button;
 - #1790 - DE translation;